Switched to flipping cards and combined some shared components of show cards and portfolio cards
